// src/common/types/remain-days.ts
export enum RemainDaysTermLabel {
  // NCD 一级
  M_1 = '1M',
  M_3 = '3M',
  M_6 = '6M',
  M_9 = '9M',
  Y_1 = '1Y',

  // 其余
  M_0_3 = '<3M',
  M_3_6 = '3-6M',
  M_6_9 = '6-9M',
  M_9_12 = '9-12M',
  Y_1_2 = '1-2Y',
  Y_2_E = '>2Y',
  Y_2_3 = '2-3Y',
  Y_3_4 = '3-4Y',
  Y_4_5 = '4-5Y',
  Y_5_10 = '5-10Y',
  Y_10_E = '>10Y'
}

// 月份
export enum RemainDaysMonthLabel {
  month_1 = '1月',
  month_2 = '2月',
  month_3 = '3月',
  month_4 = '4月',
  month_5 = '5月',
  month_6 = '6月',
  month_7 = '7月',
  month_8 = '8月',
  month_9 = '9月',
  month_10 = '10月',
  month_11 = '11月',
  month_12 = '12月'
}

// 季度
export enum RemainDaysSeasonLabel {
  Season_1 = '第一季度',
  Season_2 = '第二季度',
  Season_3 = '第三季度',
  Season_4 = '第四季度'
}